flaskProject1/requirements.txt
Flask-SQLAlchemy
python-dotenv
PyMySQL
